changes for upstream codegen changes
due to the usage of the new codegen library on latest versions of `@synthetixio/queries`, some of the subgraph queries have
changed in name/functionality a bit.

This PR implements the necessary changes into staking to be compatible with the latest version of `@synthetixio/queries`.

in hindsight, the need to supply a `subgraphUrl` is a bit cumbersome and should be read from the context like all the other query functions, but
for the time being it isn't too much of a problem.

also, there was some duplicate functionality for `useClaimedStatus.ts` which was ultimately already provided by `useUserStakingData.ts`, so I removed.

verified functionality:
* debt page works same as before
* history page works same as before
* user claim seems to work same as before

please verify me.